requirements.txt
Flask
python-dotenv
google-generativeai
google
flask-wtf
gunicorn